Fix test case when trying to share empty file
The empty test file is named ".empty" and the actual error which happens
when trying to share that file is because it doesn't have the extension
".go". So no test case is actually failing because some file is empty.

Change the empty test file name to ".empty.go" so the extension is
correct and the test can fail because the file is empty.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>